feat(apiserver): 网站分组相关

+++ apiserver/v1/ACMEUser.go
@@ -12,8 +12,8 @@ const TableNameACMEUser = "galloACMEUsers"
 // ACMEUser ACME用户
 type ACMEUser struct {
 	metav1.ObjectMeta `json:"metadata,omitempty"`
-	AdminID           int32  `gorm:"column:adminId;comment:管理员ID" json:"adminId"`                               // 管理员ID
-	UserID            int32  `gorm:"column:userId;comment:用户ID" json:"userId"`                                  // 用户ID
+	AdminID           uint64 `gorm:"column:adminId;comment:管理员ID" json:"adminId"`                               // 管理员ID
+	UserID            uint64 `gorm:"column:userId;comment:用户ID" json:"userId"`                                  // 用户ID
 	PrivateKey        string `gorm:"column:privateKey;comment:私钥" json:"privateKey"`                            // 私钥
 	Email             string `gorm:"column:email;comment:E-mail" json:"email"`                                  // E-mail
 	State             bool   `gorm:"column:state;default:1;comment:状态" json:"state"`                            // 状态
