+++ apiserver/v1/RegionProvince.go
@@ -4,9 +4,9 @@ const TableNameGalloRegionProvince = "galloRegionProvinces"
 
 // RegionProvince 区域-省份
 type RegionProvince struct {
-	ID          int32  `gorm:"column:id;primaryKey;autoIncrement:true;comment:ID" json:"id"` // ID
-	ValueID     int32  `gorm:"column:valueId;comment:实际ID" json:"valueId"`                   // 实际ID
-	CountryID   int32  `gorm:"column:countryId;comment:国家ID" json:"countryId"`               // 国家ID
+	ID          uint64 `gorm:"column:id;primaryKey;autoIncrement:true;comment:ID" json:"id"` // ID
+	ValueID     uint64 `gorm:"column:valueId;comment:实际ID" json:"valueId"`                   // 实际ID
+	CountryID   uint64 `gorm:"column:countryId;comment:国家ID" json:"countryId"`               // 国家ID
 	Name        string `gorm:"column:name;comment:名称" json:"name"`                           // 名称
 	Codes       string `gorm:"column:codes;comment:代号" json:"codes"`                         // 代号
 	CustomName  string `gorm:"column:customName;comment:自定义名称" json:"customName"`            // 自定义名称
